+++ assets/AceEditorStyleAsset.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace app\assets;
 
 use yii\web\AssetBundle;
