+++ assets/AdminLteUserAsset.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace app\assets;
+
+use yii\web\AssetBundle;
+
+/**
+ * AdminLteUserAsset asset bundle.
+ */
+class AdminLteUserAsset extends AssetBundle
+{
+    public $basePath = '@webroot';
+    public $baseUrl = '@web';
+    public $css = [
+        'css/site.css',
+        'css/adminlte-fix.css',
+    ];
+    public $js = [
+    ];
+    public $depends = [
+    ];
+}
