+++ controllers/DataController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace app\controllers;
 
 use app\models\Country;
