+++ modules/guest/Module.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace app\modules\guest;
+
+/**
+ * public module definition class
+ */
+class Module extends \yii\base\Module
+{
+    /**
+     * {@inheritdoc}
+     */
+    public $controllerNamespace = 'app\modules\guest\controllers';
+
+    /**
+     * {@inheritdoc}
+     */
+    public function init()
+    {
+        parent::init();
+        $this->layout = 'main';
+    }
+}
