+++ modules/guest/views/default/error.php
@@ -1,0 +1,28 @@
+<?php
+
+/* @var $this yii\web\View */
+/* @var $name string */
+/* @var $message string */
+
+/* @var $exception Exception */
+
+use yii\helpers\Html;
+
+$this->title = $name;
+?>
+<div class="site-error">
+
+    <h1><?= Html::encode($this->title) ?></h1>
+
+    <div class="alert alert-danger">
+        <?= nl2br(Html::encode($message)) ?>
+    </div>
+
+    <p>
+        The above error occurred while the Web server was processing your request.
+    </p>
+    <p>
+        Please <?= Html::a('contact', 'https://github.com/opensourcewebsite-org/opensourcewebsite-org/issues') ?> us if you think this is a server error. Thank you.
+    </p>
+
+</div>
