+++ views/site/account.php
@@ -100,6 +100,3 @@ $this->title = 'Account';
         </div>
     </div>
 </div>
-<?php if (!Yii::$app->user->isGuest && !Yii::$app->user->identity->is_email_confirmed): ?>
-    <?php echo Html::a('Resend Confirmation Email', ['site/resend-confirmation-email'], ['class' => 'btn btn-primary']); ?>
-<?php endif;?>
